Move files to unit test folder in designer-ui
